+++ src/Mapper.php
@@ -9,13 +9,12 @@
 namespace Blast\Orm;
 
 use ArrayObject;
-use Blast\Orm\Data\DataHelper;
 use Blast\Orm\Data\DataObject;
-use Blast\Orm\Data\DataObjectInterface;
-use Blast\Orm\Data\UpdatedDataObjectInterface;
 use Blast\Orm\Entity\EntityAdapter;
 use Blast\Orm\Entity\EntityAdapterInterface;
 use Blast\Orm\Entity\EntityAdapterLoaderTrait;
+use Blast\Orm\Entity\EntityAwareInterface;
+use Blast\Orm\Entity\EntityAwareTrait;
 use Blast\Orm\Query;
 use Blast\Orm\Query\Result;
 use stdClass;
